requirements.txt
# Azure Functions for Python - Core dependencies
azure-functions==1.18.0
azure-functions-worker==1.0.15

# Azure SDK
azure-identity==1.16.1
azure-keyvault-secrets==4.8.0
azure-storage-queue==12.9.0
azure-servicebus==7.12.0

# HTTP and API client
requests==2.31.0
httpx==0.27.0

# Logging and monitoring
opencensus-ext-azure==1.1.13
opencensus-ext-logging==0.1.1

# Data processing
pydantic==2.7.1
python-dateutil==2.9.0.post0

# Development and testing
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-mock==3.14.0

# Code quality
ruff==0.4.4
black==24.4.0
mypy==1.10.0

# FastAPI for mock services in testing
fastapi==0.111.0
uvicorn==0.29.0